fix(combat): base damage override channel carries fractions to the final round
QA: los +0.25/+0.50 de Furia Contenida no sumaban al danio final. El canal
entero era int de punta a punta y el reader ademas flooreaba — un 0.25
literal no podia sobrevivir. Ahora:

- EffectIntReader.ReadFloat virtual (default: el Read entero) — sin nueva
  jerarquia ni cambios de tipo serializado: los ~15 readers y los assets
  Odin que serializan por type name quedan intactos.
- IBaseDamageOverrideService.TryGetBaseDamage devuelve float via
  ReadFloat; ReadCleanTurnStreakScaled preserva streak x PerTurnAmount
  sin floor (Read legacy sigue flooreando).
- PlayerComboDamage: dmgBasePJ y N son float; RoundNxM(float, float) es
  el UNICO punto de redondeo (AwayFromZero intacto).
- Breakdown UI absorbe el float: FinalN/AttackBase float, reconcile por
  epsilon, formato "0.##" en counters y vuelos (enteros se ven igual).

El remark muerto de la "escala x10" se elimina — nunca existio en la
formula.

+++ Assets/Scripts/Rollgeon/UI/HUD/Breakdown/BreakdownJuice.cs
@@ -84,7 +84,7 @@ namespace Rollgeon.UI.HUD.Breakdown
         // Ciclo — llamados por el director
         // ==================================================================
 
-        public void OnSequenceStart(int finalN, float finalM, int finalTotal)
+        public void OnSequenceStart(float finalN, float finalM, int finalTotal)
         {
             _finalTotal = finalTotal;
             _finalM = finalM;
